Move "enable_when_ready" into it's own module
